save_all_on_run: Only save files that exist on disk and have dirty buffers
Note: the "save_all_on_run" option no longer saves files that don't exist
on disk.

The reason for this change is trying to save a file that doesn't exist
on disk prompts the user with a "save file" dialog, which is generally
not desired behaviour.

Maybe another option "save_all_on_run_strict" option can be added later
that will try to save even the files that don't exist on disk.